+++ lib/sbmt/strangler.rb
@@ -5,12 +5,25 @@ require "dry-monads"
 require "yabeda"
 require "faraday"
 require "faraday/net_http_persistent"
+require "flipper"
+require "concurrent"
+require "securerandom"
 require "oj"
 
 begin
   require "sentry-rails"
 rescue LoadError
   # optional dependency
+end
+
+begin
+  require "opentelemetry-instrumentation-concurrent_ruby"
+rescue LoadError
+  Warning.warn <<~WARN if Object.const_defined?(:Opentelemetry)
+    WARNING! It looks like you're using OpenTelemetry but you didn't install an instrumentation for the concurrent-ruby gem.
+    sbmt-strangler runs your code using concurrent-ruby futures in mirror mode, so this is very adviced to install
+    the instrumentation (opentelemetry-instrumentation-concurrent_ruby gem) to get traces in mirror mode!
+  WARN
 end
 
 require_relative "strangler/configurable"
@@ -25,6 +38,12 @@ require_relative "strangler/errors"
 require_relative "strangler/error_tracker"
 require_relative "strangler/logger"
 require_relative "strangler/http"
+require_relative "strangler/flipper"
+require_relative "strangler/feature_flags"
+require_relative "strangler/metric_tracker"
+require_relative "strangler/work_modes/proxy"
+require_relative "strangler/work_modes/mirror"
+require_relative "strangler/work_modes/replace"
 
 require_relative "strangler/engine"